+++ src/main/java/com/example/demo/jobmanagement/JobRequest.java
@@ -13,9 +13,9 @@ public class JobRequest {
     private String title;
     private String description;
     private String location;
-    private double experience;
+    private int experience;
     private String jobType;
     private LocalDateTime deadline;
-    private double salary;
+    private int salary;
 
 }
